COM-3853: rename folder
